Add full encircled energy table and code

+++ grizli/pipeline/photoz.py
@@ -1,5 +1,8 @@
-def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper_ix=1, apply_prior=False, beta_prior=True, get_external_photometry=False, external_limits=3, external_sys_err=0.3, external_timeout=300, sys_err=0.05, z_step=0.01, z_min=0.01, z_max=12, total_flux='flux_auto', compute_residuals=False, dummy_prior=False, extra_rf_filters=[]):
-    
+
+def apply_catalog_corrections(root, total_flux='flux_auto', auto_corr=True, get_external_photometry=False, aperture_indices='all', suffix='_apcorr', verbose=True, apply_background=True):
+    """
+    Aperture and background corrections to photometric catalog
+    """
     import os
     import eazy
     import numpy as np
@@ -7,21 +10,11 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
     from grizli import utils
     import mastquery.utils
     
-    if (os.path.exists('{0}.eazypy.self.npy'.format(root))) & (not force):
-        self = np.load('{0}.eazypy.self.npy'.format(root))[0]
-        zout = utils.read_catalog('{0}.eazypy.zout.fits'.format(root))
-        cat = utils.read_catalog('{0}_phot_apcorr.fits'.format(root))
-        return self, cat, zout
-        
-    trans = {'f098m':201, 'f105w':202, 'f110w':241, 'f125w':203, 'f140w':204, 'f160w':205, 'f435w':233, 'f438w':211, 'f606w':236, 'f625w':237, 'f814w':239, 'f702w':15, 'f555w':235, 'f350lp':339, 'f475w':212, 'f775w':238, 'f850lp':240}
-    #trans.pop('f814w')
-    
     cat = utils.read_catalog('{0}_phot.fits'.format(root))
     filters = []
     for c in cat.meta:
         if c.endswith('_ZP'):
             filters.append(c.split('_ZP')[0].lower())
-    
     
     if get_external_photometry:
         print('Get external photometry from Vizier')
@@ -39,19 +32,31 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
             print(' - External catalog FAILED')
             pass
         
-    # Total flux
-    
     # Fix: Take flux_auto when flag==0, flux otherwise
     if (total_flux == 'flux_auto_fix') & (total_flux not in cat.colnames):
         flux = cat['flux_auto']*1.
         flagged = (cat['flag'] > 0)
         flux[flagged] = cat['flux'][flagged]
         cat['flux_auto_fix'] = flux*1.
-        
+    
+    # Additional auto correction
+
     cat.meta['TOTALCOL'] = total_flux, 'Column for total flux'
+    #cat.meta['HASTOT'] = (auto_corr &  ('tot_corr' in cat.colnames), 'Catalog has full total flux')
     
     apcorr = {}
-    for i in range(5):
+    for NAPER in range(100):
+        if 'APER_{0}'.format(NAPER) not in cat.meta:
+            break
+    
+    if aperture_indices == 'all':
+        aperture_indices = range(NAPER)
+        
+    for i in aperture_indices:
+        
+        if verbose:
+            print('Compute aperture corrections: i={0}, D={1:.2f}" aperture'.format(i, cat.meta['ASEC_{0}'.format(i)]))
+            
         if 'flux_aper_{0}'.format(i) in cat.colnames:
             cat['apcorr_{0}'.format(i)] = cat[total_flux]/cat['flux_aper_{0}'.format(i)]
             for f in filters:
@@ -63,7 +68,7 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
                     
                 cat['{0}_corr_{1}'.format(f, i)] = (cat['{0}_flux_aper_{1}'.format(f, i)]-bkg)*cat['apcorr_{0}'.format(i)]
                 cat['{0}_ecorr_{1}'.format(f, i)] = cat['{0}_fluxerr_aper_{1}'.format(f, i)]*cat['apcorr_{0}'.format(i)]
-
+                
                 # mask_thresh = np.percentile(cat['{0}_mask_aper_{1}'.format(f, i)], 95)
                 aper_area = np.pi*(cat.meta['APER_{0}'.format(i)]/2)**2
                 mask_thresh = aper_area
@@ -72,6 +77,15 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
                 cat['{0}_corr_{1}'.format(f, i)][bad] = -99
                 cat['{0}_ecorr_{1}'.format(f, i)][bad] = -99
                 
+                tot_col = '{0}_tot_corr'.format(f.lower())
+                
+                if auto_corr and (tot_col in cat.colnames):   
+                    cat['{0}_tot_{1}'.format(f, i)] = cat['{0}_corr_{1}'.format(f, i)]*cat[tot_col]
+                    cat['{0}_etot_{1}'.format(f, i)] = cat['{0}_ecorr_{1}'.format(f, i)]*cat[tot_col]
+
+                    cat['{0}_tot_{1}'.format(f, i)][bad] = -99
+                    cat['{0}_etot_{1}'.format(f, i)][bad] = -99
+                                    
     cat.rename_column('number','id')
     cat['z_spec'] = cat['id']*0.-1
     
@@ -80,6 +94,9 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
                                'data/sep_catalog_junk.pkl')
                                
     if os.path.exists(morph_model):
+        if verbose:
+            print('Apply morphological validity class')
+        
         from sklearn.externals import joblib
         clf = joblib.load(morph_model)
         X = np.hstack([[cat['peak']/cat['flux'], 
@@ -96,14 +113,47 @@ def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper
     cat['dummy_err'] =  10**(-0.4*(8-23.9))
     cat['dummy_flux'] = cat[total_flux] # detection band
     
-    cat.write('{0}_phot_apcorr.fits'.format(root), overwrite=True)
+    if suffix:
+        if verbose:
+            print('Write {0}_phot{1}.fits'.format(root, suffix))
+            
+        cat.write('{0}_phot{1}.fits'.format(root, suffix), overwrite=True)
     
+    return cat
+    
+def eazy_photoz(root, force=False, object_only=True, apply_background=True, aper_ix=1, apply_prior=False, beta_prior=True, get_external_photometry=False, external_limits=3, external_sys_err=0.3, external_timeout=300, sys_err=0.05, z_step=0.01, z_min=0.01, z_max=12, total_flux='flux_auto', auto_corr=True, compute_residuals=False, dummy_prior=False, extra_rf_filters=[]):
+    
+    import os
+    import eazy
+    import numpy as np
+    
+    from grizli import utils
+    import mastquery.utils
+    
+    if (os.path.exists('{0}.eazypy.self.npy'.format(root))) & (not force):
+        self = np.load('{0}.eazypy.self.npy'.format(root))[0]
+        zout = utils.read_catalog('{0}.eazypy.zout.fits'.format(root))
+        cat = utils.read_catalog('{0}_phot_apcorr.fits'.format(root))
+        return self, cat, zout
+        
+    trans = {'f098m':201, 'f105w':202, 'f110w':241, 'f125w':203, 'f140w':204, 'f160w':205, 'f435w':233, 'f438w':211, 'f606w':236, 'f625w':237, 'f814w':239, 'f702w':15, 'f555w':235, 'f350lp':339, 'f475w':212, 'f775w':238, 'f850lp':240}
+    #trans.pop('f814w')
+    
+    print('Apply catalog corrections')
+    apply_catalog_corrections(root, suffix='_apcorr')
+    
+    cat = utils.read_catalog('{0}_phot_apcorr.fits'.format(root))
+    filters = []
+    for c in cat.meta:
+        if c.endswith('_ZP'):
+            filters.append(c.split('_ZP')[0].lower())
+        
     # Translate
     fp = open('zphot.translate','w')
     for f in filters:
         if f in trans:
-            fp.write('{0}_corr_{1} F{2}\n'.format(f, aper_ix, trans[f]))
-            fp.write('{0}_ecorr_{1} E{2}\n'.format(f, aper_ix, trans[f]))
+            fp.write('{0}_tot_{1} F{2}\n'.format(f, aper_ix, trans[f]))
+            fp.write('{0}_etot_{1} E{2}\n'.format(f, aper_ix, trans[f]))
     
     fp.write('irac_ch1_flux F18\n')
     fp.write('irac_ch1_err  E18\n')
@@ -371,6 +421,8 @@ def get_external_catalog(phot, filter_file='/usr/local/share/eazy-photoz/filters
     
 ########### Selecting objects
 def select_objects():
+    from grizli.pipeline import photoz
+    import numpy as np
     
     total_flux = 'flux_auto_fix'
     total_flux = 'flux_auto' # new segmentation masked SEP catalogs
@@ -378,7 +430,7 @@ def select_objects():
     self, cat, zout = photoz.eazy_photoz(root, object_only=False, apply_prior=False, beta_prior=True, aper_ix=1, force=True, get_external_photometry=False, compute_residuals=False, total_flux=total_flux)
 
     flux = self.cat[total_flux]*1.
-    hmag = 23.9-2.5*np.log10(flux)
+    hmag = 23.9-2.5*np.log10(cat['f160w_tot_2'])
 
     # Reddest HST band
     lc_clip = self.lc*1
@@ -396,12 +448,22 @@ def select_objects():
     iz8 = np.where(self.zgrid > 8)[0][0]
     iz9 = np.where(self.zgrid > 9)[0][0]
 
-    highz_sel = (cumpz[:,iz6] > 0) & ((cumpz[:,iz7] < 0.3) | (cumpz[:,iz8] < 0.4) |  (cumpz[:,iz8] < 0.5)) 
+    highz_sel = (hmag < 27.5) & (self.cat['class_valid'] > 0.8)
     #highz_sel |= (cumpz[:,iz6] < 0.3) & (self.cat['flux_radius'] > 2.5) 
-    highz_sel &= (hmag < 27.5) & (self.cat['class_valid'] > 0.8)
     #highz_sel &= (self.cat['flux_radius'] > 2.5) 
     highz_sel &= chi2 < 3
     highz_sel &= (sn_red > 5)
+    highz_sel &= self.nusefilt >= 3
+    
+    flux_ratio = (cat['f160w_flux_aper_3'] - cat['f160w_bkg_aper_3'])/(cat['f160w_flux_aper_0'] - cat['f160w_bkg_aper_0'])
+    flux_ratio /= cat.meta['APER_3']**2/cat.meta['APER_0']**2
+    
+    if False:
+        sel = highz_sel
+        so = np.argsort(cumpz[sel,iz7]); ids = self.cat['id'][sel][so]; i=-1
+        so = np.argsort(flux_ratio[sel])[::-1]; ids = self.cat['id'][sel][so]; i=-1
+        
+    highz_sel &= (cumpz[:,iz6] > 0) & (flux_ratio > 0.4) & ((cumpz[:,iz6] < 0.3) | (cumpz[:,iz7] < 0.3) | (((cumpz[:,iz8] < 0.4) |  (cumpz[:,iz9] < 0.5)) & (flux_ratio < 0.5))) 
 
     # Big objects likely diffraction spikes
     # big = (self.cat['flux_radius'] > 10)
@@ -414,7 +476,7 @@ def select_objects():
     # Red 
     uv = -2.5*np.log10(zout['restU']/zout['restV'])
     red_sel = ((zout['z160'] > 1.) & (uv > 1.5)) | ((zout['z160'] > 1.5) & (uv > 1.1))
-    red_sel &= (self.zbest < 4) & (hmag < 22) & (~hmag.mask) 
+    red_sel &= (self.zbest < 4) & (hmag < 22)# & (~hmag.mask) 
     red_sel &= (zout['mass'] > 10**10.5) #& (self.cat['class_valid'] > 0.8)
     red_sel &= (self.cat['flux_radius'] > 2.5)
     red_sel &= (zout['restV']/zout['restV_err'] > 3)
@@ -445,7 +507,7 @@ def select_objects():
         plt.close()
 
         # Cutout
-        from grizli_aws.aws_drizzle import drizzle_images
+        #from grizli_aws.aws_drizzle import drizzle_images
         
         #rgb_params = {'output_format': 'png', 'output_dpi': 75, 'add_labels': False, 'show_ir': False, 'suffix':'.rgb'}
         rgb_params = None
@@ -455,7 +517,7 @@ def select_objects():
 
         label = '{0}_{1:05d}'.format(root, id_j)
         if not os.path.exists('{0}.rgb.png'.format(label)):
-            drizzle_images(label=label, ra=ra, dec=dec, pixscale=0.06, size=8, pixfrac=0.8, theta=0, half_optical_pixscale=False, filts=['f160w', 'f814w', 'f140w', 'f125w', 'f105w', 'f110w', 'f098m', 'f850lp', 'f775w', 'f606w', 'f475w'], remove=False, rgb_params=rgb_params, master='grizli-jan2019', aws_bucket=aws_bucket)
+            drizzle_images(label=label, ra=ra, dec=dec, pixscale=0.06, size=8, pixfrac=0.8, theta=0, half_optical_pixscale=False, filters=['f160w', 'f814w', 'f140w', 'f125w', 'f105w', 'f110w', 'f098m', 'f850lp', 'f775w', 'f606w', 'f475w'], remove=False, rgb_params=rgb_params, master='grizli-jan2019', aws_bucket=aws_bucket)
 
         show_all_thumbnails(label=label, filters=['f775w','f814w','f098m','f105w','f110w','f125w','f140w','f160w'], scale_ab=np.clip(hmag[j],19,22), close=True)
